Changing entire file structure
