+++ src/almanack/repository_processing.py
@@ -1,117 +1,12 @@
-import json
-import pathlib
 import shutil
 import tempfile
 from datetime import datetime, timezone
 from typing import Dict, List
-
 import pygit2
 
-from .entropy import (
-    aggregate_entropy_calculation,
-    calculate_normalized_entropy,
-    get_edited_files,
-)
-
-
-def clone_repository(repo_url: str) -> pathlib.Path:
-    """
-    Clones the GitHub repository to a temporary directory.
-
-    Args:
-        repo_url (str): The URL of the GitHub repository.
-
-    Returns:
-        pathlib.Path: Path to the cloned repository.
-    """
-    temp_dir = tempfile.mkdtemp()
-    repo_path = pathlib.Path(temp_dir) / "repo"
-    pygit2.clone_repository(repo_url, str(repo_path))
-    return repo_path
-
-
-def get_commits(repo: pygit2.Repository) -> List[pygit2.Commit]:
-    """
-    Retrieves the list of commits from the main branch.
-
-    Args:
-        repo (pygit2.Repository): The Git repository.
-
-    Returns:
-        List[pygit2.Commit]: List of commits in the repository.
-    """
-    head = repo.revparse_single("HEAD")
-    walker = repo.walk(
-        head.id, pygit2.GIT_SORT_NONE
-    )  # Use GIT_SORT_NONE to ensure all commits are included
-    commits = [commit for commit in walker]
-
-    return commits
-
-
-def save_entropy_to_json(
-    repo_url: str,
-    file_level_entropy: Dict[str, float],
-    normalized_total_entropy: float,
-    output_path: str,
-) -> None:
-    """
-    Saves the entropy values and the top 5 files with the most entropy to a JSON file.
-
-    Args:
-        repo_url (str): The URL of the GitHub repository.
-        file_level_entropy (Dict[str, float]): Entropy values for each file.
-        normalized_total_entropy (float): Total normalized entropy for the repository.
-        output_path (str): Path to the output JSON file.
-    """
-    # Sort the files by entropy in descending order and get the top 5
-    sorted_files_by_entropy = sorted(
-        file_level_entropy.items(), key=lambda item: item[1], reverse=True
-    )[:5]
-
-    data = {
-        "repo_url": repo_url,
-        "total_normalized_entropy": normalized_total_entropy,
-        "file_level_entropy": file_level_entropy,
-        "5_files_with_most_entropy": sorted_files_by_entropy,
-    }
-
-    with open(output_path, "w") as json_file:
-        json.dump(data, json_file, indent=4)
-
-
-def print_entropy_report(output_path: str) -> None:
-    """
-    Prints the entropy values from a JSON file.
-
-    Args:
-        output_path (str): Path to the output JSON file.
-    """
-    with open(output_path, "r") as json_file:
-        data = json.load(json_file)
-
-    border = "=" * 50
-    separator = "-" * 50
-    title = "Entropy Report"
-
-    print(border)
-    print(f"{title:^50}")  # Centers the title
-    print(border)
-
-    # Print repository details
-    print(f"Repository URL: {data['repo_url']}")
-    print(f"Total Normalized Entropy: {data['total_normalized_entropy']:.4f}")
-
-    # Print top 5 files with most entropy
-    print("\nTop 5 Files with Most Entropy:")
-    print(separator)
-    for file_name, entropy in data["5_files_with_most_entropy"]:
-        print(f"  {file_name:<40} {entropy:>8.4f}")
-    print(separator)
-
-    # Print the report footer
-    print(border)
-
+from .git_operations import clone_repository, get_commits, get_edited_files
+from .entropy import calculate_normalized_entropy, aggregate_entropy_calculation
+from .data_management import save_entropy_to_json, print_entropy_report
 
 def process_repository(repo_url: str, output_path: str) -> None:
     """
