+++ tests/test_entropy.py
@@ -1,13 +1,12 @@
 """
-Testing entropy.py functionality
+Testing entropy functionality
 """
+
 
 import pathlib
 
-from test_git_parser import get_most_recent_commits
-
 from almanack.entropy import aggregate_entropy_calculation, calculate_normalized_entropy
-
+from .test_git_operations import get_most_recent_commits
 
 def test_calculate_normalized_entropy(
     repository_paths: dict[str, pathlib.Path], repo_file_sets: dict[str, list[str]]
@@ -32,7 +31,7 @@ def test_calculate_normalized_entropy(
             )  # Check if entropy is non-negative and within normalized range of [0,1]
 
 
-def test_aggregate_entropy_different_repos(
+def test_aggregate_entropy_calculation(
     repository_paths: dict[str, pathlib.Path], repo_file_sets: dict[str, list[str]]
 ) -> None:
     """
