+++ tests/test_repository_processing.py
@@ -1,7 +1,13 @@
-from almanack.repo_helper import process_repository
+"""
+Testing repo_helper functionality
+"""
 
+from almanack.repository_processing import process_repository
 
-def test_repo_helper():
+def test_repo_helper() -> None:
+    """
+    Test process_repository function.
+    """
     output_path = (
         "/home/willdavidson/Desktop/Almanack/almanac/tests/entropy_report.json"
     )
